HeroList & HeroListPage結構簡化
