--- src/components/HeroList/HeroList.type.ts
@@ -1,7 +1,0 @@
-import { Hero } from "src/types/hero.type";
-
-export interface HeroListProps {
-  heros: Hero[];
-  selectedIndex?: number;
-  onCardClick?: (index: number) => void;
-}
