+++ src/pages/HeroList/HeroList.style.ts
@@ -1,13 +1,12 @@
-import { $Rwd } from 'src/style/rwd';
-import styled from 'styled-components';
+import { $Rwd } from "src/style/rwd";
+import styled from "styled-components";
 
-export const StyledHeroList = styled.div`
+export const ListBlock = styled.div`
 	display: grid;
 	gap: 10px;
 	grid-template-columns: 1fr 1fr 1fr 1fr;
-	
+
 	@media (max-width: ${$Rwd.tablet}) {  
 		grid-template-columns: 1fr 1fr;
 	}
-}
 `;
