+++ src/pages/HeroList/HeroList.type.ts
@@ -1,3 +1,7 @@
+export interface HeroListPageProps {
+
+}
+
 export interface Hero {
 	id: string;
 	name: string;
