--- src/pages/HeroListPage/HeroListPage.type.ts
@@ -1,3 +1,0 @@
-export interface HeroListPageProps {
-
-}
