+++ src/pages/HeroProfile/HeroFrofile.service.ts
@@ -1,5 +1,4 @@
 import Api from "src/service/api";
-import { Hero } from "src/types/hero.type";
 import { Profile } from "./HeroProfile.type";
 
 export const HeroProfileService = () => {
@@ -8,7 +7,7 @@ export const HeroProfileService = () => {
 
 	return {
 		getHeroProfile: (id: string) => {
-			return api.getJSON<Hero>(`https://hahow-recruit.herokuapp.com/heroes/${id}/profile`)
+			return api.getJSON<Profile>(`https://hahow-recruit.herokuapp.com/heroes/${id}/profile`)
 		},
 		updateHeroProfile: (id: string, profile: Profile) => {
 			return api.patch(`https://hahow-recruit.herokuapp.com/heroes/${id}/profile`, profile)
